% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastICAparallel.R
\name{fastICAparallel}
\alias{fastICAparallel}
\title{fastICA implementation using a parallel scheme}
\usage{
fastICAparallel(data.mat, p, max = 200)
}
\arguments{
\item{data.mat}{white data matrix}

\item{p}{number of components}

\item{max}{maximum number of iterations}
}
\value{
ica.arr independent component array
}
\description{
Implementation of the fixed point iteration method where a parallel scheme is used to find ICs.
}
\examples{
ica <- fastICAparallel(data,3,200)

}
