% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.analysis.R
\name{ica.analysis}
\alias{ica.analysis}
\title{function to whiten data before performing ICA analysis.}
\usage{
ica.analysis(data, p, scheme = c("deflation", "parallel"), max = 200)
}
\arguments{
\item{data}{data matrix to be whiten}

\item{p}{number of components}
}
\value{
white data for ICA analysis
}
\description{
Whitening method that uses singular value decomposition.
}
\examples{
data.white <- whiten(data,3)

}
