% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomax.R
\name{infomax}
\alias{infomax}
\title{Infomax implementation}
\usage{
infomax(data.mat, p, max = 200)
}
\arguments{
\item{data.mat}{data matrix where the columns are bands}

\item{p}{number of components}

\item{max}{maximum number of iterations}
}
\value{
ica.arr independent component array
}
\description{
Implementation of the infomax algorithm to find ICs.
}
\examples{
ica <- infomax(data,3,200)

}
