% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.array.R
\name{make.array}
\alias{make.array}
\title{Function converting 80000xbands matrix to 200x400xbands array.}
\usage{
make.array(mat, rows, cols, bands)
}
\arguments{
\item{mat}{matrix with 80000xbands dimension}

\item{bands}{number of bands}
}
\value{
arr array with 200x400xbands dimension
}
\description{
Function converting 80000xbands matrix to 200x400xbands array.
}
\examples{
arr <- make.array(mat,bands)

}
