% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dataCube.R
\name{read.dataCube}
\alias{read.dataCube}
\title{Function for creating data matrix from raw data file.}
\usage{
read.dataCube(HSIdata, rows = 200, cols = 400, bands = 216)
}
\arguments{
\item{HSIdata}{csv file}

\item{rows}{number of rows}

\item{cols}{number of columns}

\item{bands}{number of bands}
}
\value{
data.mat data in matrix format
}
\description{
Creates a data matrix from a .csv file containing HSI data. The input data must be (rowsxcols)xbands
}
\examples{
data.mat <- read.dataCube("syntetisk_aeble.csv",rows,cols,bands)

}
