% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Master.thesis.R
\docType{package}
\name{thesis.package}
\alias{thesis.package}
\alias{thesis.package-package}
\title{Master thesis Package}
\description{
Implementation of code use din the master thesis to find the components in HSI data.
}
\author{
Nina Louise Pedersen \email{npede14@student.sdu.dk}
}
