% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastICAdeflation.R
\name{fastICAdeflation}
\alias{fastICAdeflation}
\title{fastICA implementation using a deflation scheme}
\usage{
fastICAdeflation(data.mat, p, iter = 200)
}
\arguments{
\item{data.mat}{white data matrix}

\item{p}{number of components}

\item{iter}{number of iterations}
}
\value{
ica.arr independent component array
}
\description{
Implementation of the fixed point iteration method where a deflation scheme is used to find ICs.
}
\examples{
ica <- fastICAdeflation(data,3,200)

}
