% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.analysis.R
\name{ica.analysis}
\alias{ica.analysis}
\title{Function to perform the fastICA.}
\usage{
ica.analysis(data, p, scheme = c("deflation", "parallel"), iter = 200)
}
\arguments{
\item{data}{data matrix to be whiten}

\item{p}{number of components}

\item{scheme}{choose the method for the fastICA}

\item{iter}{number of iterations}
}
\value{
white data for ICA analysis
}
\description{
Function that estimates the independent components.
}
\examples{
data.white <- ica.analysis(data,3)

}
