% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dataCube.R
\name{read.dataCube}
\alias{read.dataCube}
\title{Function for creating data matrix from raw data file.}
\usage{
read.dataCube(HSIdata, rows)
}
\arguments{
\item{HSIdata}{csv file}

\item{rows}{number of rows}
}
\value{
data.mat data in matrix format
}
\description{
Implementation of the fixed point iteration method where a deflation scheme is used to find ICs.
}
\examples{
data.mat <- read.dataCube("syntetisk_aeble.csv",rows)

}
